added if != null and @Nullable, @NotNull

+++ src/main/java/ru/burmistrov/tm/service/ProjectService.java
@@ -1,5 +1,7 @@
 package ru.burmistrov.tm.service;
 
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import ru.burmistrov.tm.api.repository.IProjectRepository;
 import ru.burmistrov.tm.api.repository.ITaskRepository;
 import ru.burmistrov.tm.api.service.IProjectService;
@@ -11,66 +13,80 @@ import java.util.List;
 
 public final class ProjectService extends AbstractService implements IProjectService {
 
+    @Nullable
     private final IProjectRepository<AbstractEntity> projectRepository;
 
+    @Nullable
     private final ITaskRepository<AbstractEntity> taskRepository;
 
-    public ProjectService(IProjectRepository<AbstractEntity> projectRepository, ITaskRepository<AbstractEntity> taskRepository) {
+    public ProjectService(@Nullable IProjectRepository<AbstractEntity> projectRepository,@Nullable ITaskRepository<AbstractEntity> taskRepository) {
         this.projectRepository = projectRepository;
         this.taskRepository = taskRepository;
     }
 
-    public void remove(String userId, String projectId) {
+    public void remove(@Nullable String userId,@Nullable String projectId) {
             Project project = new Project();
             project.setUserId(userId);
             project.setId(projectId);
-            AbstractEntity abstractEntity = projectRepository.findOne(project);
-            if(abstractEntity != null) {
-                projectRepository.remove(project);
-                Task task = new Task();
-                task.setUserId(userId);
-                task.setProjectId(projectId);
-                taskRepository.removeAllInProject(task);
+            if(projectRepository != null && taskRepository != null) {
+                AbstractEntity abstractEntity = projectRepository.findOne(project);
+                if (abstractEntity != null) {
+                    projectRepository.remove(project);
+                    Task task = new Task();
+                    task.setUserId(userId);
+                    task.setProjectId(projectId);
+                    taskRepository.removeAllInProject(task);
+                }
             }
 
     }
 
-    public Project persist(String userId, String name, String description) {
+    public Project persist(@Nullable String userId,@Nullable String name,@Nullable String description) {
 
         Project project = new Project();
         project.setUserId(userId);
         project.setName(name);
         project.setDescription(description);
-        AbstractEntity abstractEntity = projectRepository.findOne(project);
-        if(abstractEntity == null) {
-            return (Project) projectRepository.persist(project);
+        if(projectRepository != null) {
+            AbstractEntity abstractEntity = projectRepository.findOne(project);
+            if (abstractEntity == null) {
+                return (Project) projectRepository.persist(project);
+            }
         }
         return null;
     }
 
-    public void merge(String userId, String projectId, String name, String description) {
+    public void merge(@Nullable String userId,@Nullable String projectId,@Nullable String name,@Nullable String description) {
         Project project = new Project();
         project.setId(projectId);
         project.setUserId(userId);
         project.setName(name);
         project.setDescription(description);
-        AbstractEntity abstractEntity = projectRepository.findOne(project);
-        if (abstractEntity != null) {
-            projectRepository.merge(project);
+        if(projectRepository != null) {
+            AbstractEntity abstractEntity = projectRepository.findOne(project);
+            if (abstractEntity != null) {
+                projectRepository.merge(project);
+            }
         }
     }
 
-    public void removeAll(String userId) {
+    public void removeAll(@Nullable String userId) {
         Project project = new Project();
         project.setUserId(userId);
-        projectRepository.removeAll(project);
-        taskRepository.removeAll(project);
+        if(projectRepository != null && taskRepository != null) {
+            projectRepository.removeAll(project);
+            taskRepository.removeAll(project);
+        }
     }
 
+    @Nullable
     public List<AbstractEntity> findAll(String userId) {
         Project project = new Project();
         project.setUserId(userId);
-        return projectRepository.findAll(project);
+        if (projectRepository != null) {
+            return projectRepository.findAll(project);
+        }
+        return null;
     }
 
 }
