+++ src/main/java/ru/burmistrov/tm/service/TaskService.java
@@ -1,5 +1,7 @@
 package ru.burmistrov.tm.service;
 
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import ru.burmistrov.tm.api.repository.ITaskRepository;
 import ru.burmistrov.tm.api.service.ITaskService;
 import ru.burmistrov.tm.entity.AbstractEntity;
@@ -9,52 +11,61 @@ import java.util.List;
 
 public final class TaskService extends AbstractService implements ITaskService {
 
-    private final ITaskRepository<AbstractEntity> taskRepository;
+    @Nullable private final ITaskRepository<AbstractEntity> taskRepository;
 
-    public TaskService(ITaskRepository<AbstractEntity> taskRepository) {
+    public TaskService(@Nullable ITaskRepository<AbstractEntity> taskRepository) {
         this.taskRepository = taskRepository;
     }
 
-    public Task persist(String userId, String projectId, String name, String description) {
+    @Nullable
+    public Task persist(@Nullable String userId,@Nullable String projectId,@Nullable String name,@Nullable String description) {
         Task task = new Task();
         task.setName(name);
         task.setDescription(description);
         task.setProjectId(projectId);
         task.setUserId(userId);
-        AbstractEntity abstractEntity = taskRepository.findOne(task);
-        if(name.length() != 0 && abstractEntity == null) {
-
-            return (Task) taskRepository.persist(task);
+        if(taskRepository != null) {
+            AbstractEntity abstractEntity = taskRepository.findOne(task);
+            if (name != null && name.length() != 0 && abstractEntity == null) {
+                return (Task) taskRepository.persist(task);
+            }
         }
         return null;
     }
 
-    public void merge(String userId, String projectId, String taskId, String newName, String description) {
+    public void merge(@Nullable String userId, @Nullable String projectId,@Nullable String taskId,@Nullable String newName,@Nullable String description) {
         Task task = new Task();
         task.setId(taskId);
         task.setName(newName);
         task.setDescription(description);
         task.setProjectId(projectId);
         task.setUserId(userId);
-        AbstractEntity abstractEntity = taskRepository.findOne(task);
-
-        if (newName.length() != 0 && abstractEntity != null) {
-            taskRepository.merge(task);
+        if (taskRepository != null) {
+            AbstractEntity abstractEntity = taskRepository.findOne(task);
+            if (newName != null && newName.length() != 0 && abstractEntity != null) {
+                taskRepository.merge(task);
+            }
         }
     }
 
-    public List<AbstractEntity> findAll(String userId, String projectId) {
+    @Nullable
+    public List<AbstractEntity> findAll(@Nullable String userId,@Nullable String projectId) {
         Task task = new Task();
         task.setUserId(userId);
         task.setProjectId(projectId);
-        return taskRepository.findAll(task);
+        if (taskRepository != null) {
+            return taskRepository.findAll(task);
+        }
+        return null;
     }
 
     public void removeAllInProject(String userId, String projectId) {
         Task task = new Task();
         task.setUserId(userId);
         task.setProjectId(projectId);
-        taskRepository.removeAllInProject(task);
+        if (taskRepository != null) {
+            taskRepository.removeAllInProject(task);
+        }
     }
 
     public void remove(String userId, String projectId, String taskId) {
@@ -62,12 +73,16 @@ public final class TaskService extends AbstractService implements ITaskService {
         task.setProjectId(userId);
         task.setProjectId(projectId);
         task.setId(taskId);
-        taskRepository.remove(task);
+        if (taskRepository != null) {
+            taskRepository.remove(task);
+        }
     }
 
-    public void removeAll(String userId) {
+    public void removeAll(@Nullable String userId) {
         Task task = new Task();
         task.setUserId(userId);
-        taskRepository.removeAll(task);
+        if (taskRepository != null) {
+            taskRepository.removeAll(task);
+        }
     }
 }
