+++ src/main/java/ru/burmistrov/tm/service/UserService.java
@@ -1,5 +1,6 @@
 package ru.burmistrov.tm.service;
 
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import ru.burmistrov.tm.api.repository.IUserRepository;
 import ru.burmistrov.tm.api.service.IUserService;
@@ -9,18 +10,24 @@ import ru.burmistrov.tm.entity.User;
 
 public final class UserService extends AbstractService implements IUserService {
 
-    private final IUserRepository<AbstractEntity> userRepository;
+    @Nullable private final IUserRepository<AbstractEntity> userRepository;
 
-    public UserService(IUserRepository<AbstractEntity> userRepository) {
+    public UserService(@Nullable IUserRepository<AbstractEntity> userRepository) {
         this.userRepository = userRepository;
     }
 
     @Nullable
-    public User logIn(@Nullable String login,@Nullable String auth) {
-        return userRepository.logIn(login, auth);
+    public User logIn(@Nullable String login,@Nullable String auth)
+    {
+        if (userRepository != null) {
+            return userRepository.logIn(login, auth);
+        }
+        return null;
     }
 
-    public User persist(String login, String password, String firstName, String middleName, String lastName, String email, Role roleType) {
+    @Nullable
+    public User persist(@Nullable String login,@Nullable String password,@Nullable String firstName,@Nullable String middleName,@Nullable String lastName,
+                        @Nullable String email,@Nullable Role roleType) {
 
         User user = new User();
         user.setRole(roleType);
@@ -30,20 +37,23 @@ public final class UserService extends AbstractService implements IUserService {
         user.setMiddleName(middleName);
         user.setLastName(lastName);
         user.setEmail(email);
-        AbstractEntity abstractEntity = userRepository.findOne(user);
-        if(abstractEntity == null) {
-            return (User) userRepository.persist(user);
+        if (userRepository != null) {
+            AbstractEntity abstractEntity = userRepository.findOne(user);
+            if (abstractEntity == null) {
+                return (User) userRepository.persist(user);
+            }
         }
         return null;
     }
 
-    public void updatePassword(String userId, String login, String password) {
-        if (password.length() > 0) {
+    public void updatePassword(@Nullable String userId,@Nullable String login,@Nullable String password) {
+        if (password != null && password.length() > 0 && userRepository != null) {
             userRepository.updatePassword(userId, login, password);
         }
     }
 
-    public void merge(String userId, String firstName, String middleName, String lastName, String email, Role role, String login) {
+    public void merge(@Nullable String userId, @Nullable String firstName, @Nullable String middleName, @Nullable String lastName, @Nullable String email,
+                       @Nullable Role role, @Nullable String login) {
         User currentUser = new User();
         currentUser.setFirstName(firstName);
         currentUser.setMiddleName(middleName);
@@ -52,22 +62,28 @@ public final class UserService extends AbstractService implements IUserService {
         currentUser.setId(userId);
         currentUser.setRole(role);
         currentUser.setLogin(login);
-        AbstractEntity abstractEntity = userRepository.findOne(currentUser);
-        if (abstractEntity != null)
-            userRepository.merge(currentUser);
+        if (userRepository != null) {
+            AbstractEntity abstractEntity = userRepository.findOne(currentUser);
+            if (abstractEntity != null)
+                userRepository.merge(currentUser);
+        }
     }
 
     @Override
-    public void remove(String userId) {
+    public void remove(@Nullable String userId) {
         User user = new User();
         user.setId(userId);
-        userRepository.remove(user);
+        if (userRepository != null) {
+            userRepository.remove(user);
+        }
     }
 
     @Override
-    public void removeAll(String userId) {
+    public void removeAll(@Nullable String userId) {
         User user = new User();
         user.setId(userId);
-        userRepository.removeAll(user);
+        if (userRepository != null) {
+            userRepository.removeAll(user);
+        }
     }
 }
